Convert for in-browser use
This commit converts the three file fragments for use in the browser. The
CSS and JS files can be used as-is. The HTML file required a basic page
template wrapping around the contents, and some additional library
imports. In particular, the JS file is now being precompiled by
babel-standalone in the browser.