Embolden the selected history item
This commit emboldens whichever history item is the one being actively
displayed. It requires passing down as an additional prop the selected
step. The history component can then compare this reference against each
item's index during the map. Each history item is, in turn, given a
trivial boolean flag. If true, a new class is added to the item. That
class, in the CSS file, has been assigned to embolden the text.

Note that this is added to enable emboldening of a selection. Since the
latest move made starts out as the selected step when first made,
highlighting of the current move comes for free.