Add a toggle that reverses the history order
This commit adds a toggle to the bottom of the list of moves. This comes
in the form of an anchor with a click handler, and an empty `href`.  When
clicked, it causes the history to reverse the order it renders the history
items in. This is accomplished by means of a new state variable that
denotes whether the history is listing in ascending (default and previous
behaviour) or descending order. When the toggle activates, it inverts this
variable.

Since this state can be kept local to the history, it has not been pulled
up into the `Game`. Instead, the `History` now has its own constructor to
set a default value for the sorting. Then, when rendering the history
items, the order is checked. Since the JavaScript
`Array.prototype.reverse` function modifies an aray in-place, the helper
function now uses a slice of the history, so that if it needs to be
reversed, the original will not be affected. If the array is reversed, the
step number also needs to be offset - note the additional `-1` to ensure
the "Game start" correctly lines up with a step number of `0`.