Highlight the winning squares
This commit ensures the three squares responsible for winning the game are
highlighted. The core of the change is in the `calculateWinner` function,
which now returns not only the player who has won, but also the indices of
the squares that determined that result. This "winning line" and the
winner are both then tracked on the game state.

The winner state, initially undefined since it will be defined when
needed, is used in place of recalculating the winner during rendering of
the game. Note that to avoid preventing players playing once a history has
been overwritten, the check for whether to return early from the board
click handler had to expand to include a check that the currently selected
step is the latest. This also allows the winner to continue to be printed
even if the history is rolled back - until the history is overwritten that
is.

The winning line is passed down to the board. Each square is then told
whether it is in the winning line by checking if its index is in the array
- a reduction is used to avoid an explicit loop. If a square is in the
winning line, it gets a new class. The new class currently applies an
animation that changes the font colour to gold and back again. Gold is a
fairly universal winner's colour after all.

Note that there have also been some minor changes to variable declaration
in the squares' click handlers, to try and keep the definition order
logical, and minimise unneeded definitions.